Calculate ddCT from dCT data.

Calculate delta delta CT from delta CT.

calculate_ddct(dCT, factors, reference, metadata = NULL, genes = NULL)

Arguments

dCT
Imported dCT qPCR data as a data.table (i.e., output of read_dct()).

factors
A character vector listing factors for statistical comparison corresponding to a subset of the columns of each data.table object.

reference
A character vector of the same length as factors, listing the level of each factor corresponding to the reference group.

metadata
Columns in the dCT data.table which are not genes, but are not relevant for comparison. Either specify metadata or genes.

genes
A character vector listing genes corresponding to a subset of the columns of each data.table object.

Value

ddCT data as a data.table.

Author

Ayush Noori